requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
cffi==1.11.5
colorama==0.4.0
cryptography==2.3.1
cx-Freeze==5.1.1
future==0.17.0
googleapis-common-protos==1.5.3
grpcio==1.16.0
grpcio-tools==1.16.0
idna==2.7
macholib==1.11
more-itertools==4.3.0
mss==3.3.1
netifaces==0.10.7
paramiko==2.4.2
pefile==2018.8.8
Pillow==5.3.0
pluggy==0.8.0
protobuf==3.6.1
psutil==5.4.7
py==1.7.0
py-cpuinfo==4.0.0
pyasn1==0.4.4
PyAutoGUI==0.9.38
pycparser==2.19
pygame==1.9.4
PyInstaller==3.4
PyMsgBox==1.0.6
PyNaCl==1.3.0
pynput==1.4
Pypubsub==4.0.0
PyScreeze==0.1.18
pysftp==0.2.9
pyspectator==1.1.2
pytest==3.9.3
PyTweening==1.0.3
pyvalid==0.9.2
pywin32==224
pywin32-ctypes==0.2.0
six==1.11.0
uptime==3.0.1
WMI==1.4.9
wxPython==4.0.1
